updated tests now that random number generator has moved from population initialisation to microsim model
